Rename main section folder
