+++ fe/src/components/app/App.tsx
@@ -2,10 +2,10 @@ import React from 'react'
 import { StylesProvider } from '@material-ui/core/styles'
 import { Provider } from 'react-redux'
 
-import Toolbar from '../main-sections/Toolbar'
-import QuickBar from '../main-sections/QuickBar'
-import ProgressBoard from '../main-sections/ProgressBoard'
-import MissionBoard from '../main-sections/MissionBoard'
+import Toolbar from '../main/Toolbar'
+import QuickBar from '../main/QuickBar'
+import ProgressBoard from '../main/ProgressBoard'
+import MissionBoard from '../main/MissionBoard'
 import store from '../../store'
 
 import styles from './styles.module.scss'
