+++ fe/src/components/main/MissionBoard/index.ts
