+++ fe/src/components/main/MissionBoard/styles.module.scss
