+++ fe/src/components/main/ProgressBoard/index.ts
