+++ fe/src/components/main/ProgressBoard/styles.module.scss
