+++ fe/src/components/main/QuickBar/QuickBar.tsx
