+++ fe/src/components/main/QuickBar/QuickBtnSet/BtnDialog/index.ts
