+++ fe/src/components/main/QuickBar/QuickBtnSet/HotkeyHint/HotkeyHint.tsx
