+++ fe/src/components/main/QuickBar/QuickBtnSet/HotkeyHint/index.ts
