+++ fe/src/components/main/QuickBar/QuickBtnSet/HotkeyHint/styles.module.scss
