+++ fe/src/components/main/QuickBar/QuickBtnSet/QuickBtn/styles.module.scss
