+++ fe/src/components/main/QuickBar/QuickBtnSet/styles.module.scss
