+++ fe/src/components/main/QuickBar/actions.ts
