+++ fe/src/components/main/QuickBar/constants.ts
