+++ fe/src/components/main/QuickBar/interfaces.ts
