+++ fe/src/components/main/QuickBar/reducer.ts
