+++ fe/src/components/main/QuickBar/styles.module.scss
