+++ fe/src/store/interfaces.ts
@@ -1,4 +1,4 @@
-import { QuickBarState } from '../components/main-sections/QuickBar/interfaces'
+import { QuickBarState } from '../components/main/QuickBar/interfaces'
 
 export interface StoreState {
   quickBar: QuickBarState
